requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
pandas>=2.0.0,<3.0.0
ortools>=9.10.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchainhub
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
streamlit==1.54.0
openai>=1.0.0
langchain-openai>=0.2.0
langchain-ollama>=0.3.0
langsmith
langchain-google-community[calendar]>=3.0.0
tavily-python>=0.7.0
deepagents>=0.4.0
